+++ elasticsearch5/search-index/src/main/java/org/vertexium/elasticsearch5/bulk/BulkUpdateServiceConfiguration.java
@@ -7,10 +7,11 @@ public class BulkUpdateServiceConfiguration {
     public static final int POOL_SIZE_DEFAULT = 10;
     public static final int BACKLOG_SIZE_DEFAULT = 100;
     public static final Duration BULK_REQUEST_TIMEOUT_DEFAULT = Duration.ofMinutes(30);
-    public static final int MAX_BATCH_SIZE_DEFAULT = 100;
+    public static final int MAX_BATCH_SIZE_DEFAULT = 1000;
     public static final int MAX_BATCH_SIZE_IN_BYTES_DEFAULT = 10 * 1024 * 1024;
     public static final Duration BATCH_WINDOW_TIME_DEFAULT = Duration.ofMillis(1000);
     public static final int MAX_FAIL_COUNT_DEFAULT = 10;
+    public static final Integer LOG_REQUEST_SIZE_LIMIT_DEFAULT = null;
     private int poolSize = POOL_SIZE_DEFAULT;
     private int backlogSize = BACKLOG_SIZE_DEFAULT;
     private Duration bulkRequestTimeout = BULK_REQUEST_TIMEOUT_DEFAULT;
@@ -18,6 +19,7 @@ public class BulkUpdateServiceConfiguration {
     private int maxBatchSizeInBytes = MAX_BATCH_SIZE_IN_BYTES_DEFAULT;
     private Duration batchWindowTime = BATCH_WINDOW_TIME_DEFAULT;
     private int maxFailCount = MAX_FAIL_COUNT_DEFAULT;
+    private Integer logRequestSizeLimit = LOG_REQUEST_SIZE_LIMIT_DEFAULT;
 
     public int getPoolSize() {
         return poolSize;
@@ -79,6 +81,15 @@ public class BulkUpdateServiceConfiguration {
 
     public BulkUpdateServiceConfiguration setBatchWindowTime(Duration batchWindowTime) {
         this.batchWindowTime = batchWindowTime;
+        return this;
+    }
+
+    public Integer getLogRequestSizeLimit() {
+        return logRequestSizeLimit;
+    }
+
+    public BulkUpdateServiceConfiguration setLogRequestSizeLimit(Integer logRequestSizeLimit) {
+        this.logRequestSizeLimit = logRequestSizeLimit;
         return this;
     }
 }