+++ elasticsearch5/search-index/src/main/java/org/vertexium/elasticsearch5/bulk/DeleteBulkItem.java
@@ -1,26 +1,30 @@
 package org.vertexium.elasticsearch5.bulk;
 
-import org.elasticsearch.action.delete.DeleteRequest;
 import org.vertexium.ElementId;
 
 public class DeleteBulkItem extends BulkItem {
-    private final String docId;
+    private final int size;
 
     public DeleteBulkItem(
-        ElementId elementId,
+        String indexName,
+        String type,
         String docId,
-        DeleteRequest deleteRequest
+        ElementId elementId
     ) {
-        super(deleteRequest.index(), elementId, deleteRequest);
-        this.docId = docId;
+        super(indexName, type, docId, elementId);
+        size = getIndexName().length()
+            + getType().length()
+            + getDocumentId().length()
+            + getElementId().getId().length();
     }
 
-    public String getDocId() {
-        return docId;
+    @Override
+    public int getSize() {
+        return size;
     }
 
     @Override
     public String toString() {
-        return String.format("DeleteBulkItem {docId='%s'}", docId);
+        return String.format("DeleteBulkItem {docId='%s'}", getDocumentId());
     }
 }