+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # v4.9.2
+* Changed: Elasticsearch bulk to squash multiple updates to same element into a single update
 * Fixed: Elasticsearch 7 query string transformation was improperly stripping escape characters from fields that could not be found in the search index.
 
 # v4.9.1